Factor out diff handling code
